Move PrecursorTree to general SpectrumWriter
