+++ Writer/ParquetSpectrumWriter.cs
@@ -1,15 +1,11 @@
 ﻿using System;
 using System.Collections.Generic;
-using System.Linq;
 using System.Reflection;
-using System.Text.RegularExpressions;
 using log4net;
 using Parquet.Serialization;
-using ThermoFisher.CommonCore.Data;
 using ThermoFisher.CommonCore.Data.Business;
 using ThermoFisher.CommonCore.Data.FilterEnums;
 using ThermoFisher.CommonCore.Data.Interfaces;
-using ThermoRawFileParser.Writer.MzML;
 
 namespace ThermoRawFileParser.Writer
 {
@@ -40,18 +36,9 @@ namespace ThermoRawFileParser.Writer
         private static readonly ILog Log =
             LogManager.GetLogger(MethodBase.GetCurrentMethod().DeclaringType);
 
-        private readonly Regex _filterStringIsolationMzPattern = new Regex(@"ms\d+ (.+?) \[");
-
-        // Precursor scan number (value) and isolation m/z (key) for reference in the precursor element of an MSn spectrum
-        private readonly Dictionary<string, int> _precursorScanNumbers = new Dictionary<string, int>();
-
-        //Precursor information for scans
-        private Dictionary<int, PrecursorInfo> _precursorTree = new Dictionary<int, PrecursorInfo>();
-
         public ParquetSpectrumWriter(ParseInput parseInput) : base(parseInput)
         {
-            _precursorScanNumbers[""] = -1;
-            _precursorTree[-1] = new PrecursorInfo();
+            //nothing to do here
         }
 
         public override void Write(IRawDataPlus raw, int firstScanNumber, int lastScanNumber)
@@ -288,69 +275,6 @@ namespace ThermoRawFileParser.Writer
                 m.precursor_charge = (uint?)trailer_charge;
                 m.ion_mobility = (float?)FAIMSCV;
                 data.Add(m);
-            }
-        }
-
-        private int GetParentFromScanString(string scanString)
-        {
-            var parts = Regex.Split(scanString, " ");
-
-            //find the position of the first (from the end) precursor with a different mass 
-            //to account for possible supplementary activations written in the filter
-            var lastIonMass = parts.Last().Split('@').First();
-            int last = parts.Length;
-            while (last > 0 &&
-                   parts[last - 1].Split('@').First() == lastIonMass)
-            {
-                last--;
-            }
-
-            string parentFilter = String.Join(" ", parts.Take(last));
-            if (_precursorScanNumbers.ContainsKey(parentFilter))
-            {
-                return _precursorScanNumbers[parentFilter];
-            }
-
-            return -2; //unsuccessful parsing
-        }
-
-        
-        private int FindLastReaction(IScanEvent scanEvent, int msLevel)
-        {
-            int lastReactionIndex = msLevel - 2;
-
-            //iteratively trying find the last available index for reaction
-            while (true)
-            {
-                try
-                {
-                    scanEvent.GetReaction(lastReactionIndex + 1);
-                }
-                catch (ArgumentOutOfRangeException)
-                {
-                    //stop trying
-                    break;
-                }
-
-                lastReactionIndex++;
-            }
-
-            //supplemental activation flag is on -> one of the levels (not necissirily the last one) used supplemental activation
-            //check last two activations
-            if (scanEvent.SupplementalActivation == TriState.On)
-            {
-                var lastActivation = scanEvent.GetReaction(lastReactionIndex).ActivationType;
-                var beforeLastActivation = scanEvent.GetReaction(lastReactionIndex - 1).ActivationType;
-
-                if ((beforeLastActivation == ActivationType.ElectronTransferDissociation || beforeLastActivation == ActivationType.ElectronCaptureDissociation) &&
-                    (lastActivation == ActivationType.CollisionInducedDissociation || lastActivation == ActivationType.HigherEnergyCollisionalDissociation))
-                    return lastReactionIndex - 1; //ETD or ECD followed by HCD or CID -> supplemental activation in the last level (move the last reaction one step back)
-                else
-                    return lastReactionIndex;
-            }
-            else //just use the last one
-            {
-                return lastReactionIndex;
             }
         }
